+++ src/main/java/com/server/crews/applicant/util/AnswerMapper.java
@@ -1,4 +1,4 @@
-package com.server.crews.applicant.mapper;
+package com.server.crews.applicant.util;
 
 import com.server.crews.applicant.domain.NarrativeAnswer;
 import com.server.crews.applicant.domain.SelectiveAnswer;
